refactor: remove ProdutoUC
