+++ src/app/controller/produto/ProdutoController.js
@@ -1,34 +1,41 @@
+const UUIDGenerator = require('../../util/UUIDGenerator')
+
 class ProdutoController {
+  
   constructor(app) {
-    this.casoUso = new app.src.app.usecase.ProdutoUC(app)
+    this.app = app
+    this.repository = new this.app.src.app.repository.ProdutoRepository(this.app)
   }
 
-  async index (attributes, filter, order) {
-    return await this.casoUso.index(attributes, filter, order)
+  async index(attributes, filter, order){
+    return await this.repository.findAll(attributes, filter, order)
   }
 
   async findAndPaginate (attributes, filter, order, page) {
-    return await this.casoUso.findAndPaginate(attributes, filter, order, page)
+    return await this.repository.findAndPaginate(attributes, filter, order, page)
   }
   
   async findByUUId (UUId) {
-    return await this.casoUso.findByUUId(UUId)
+    return await this.repository.findByUUId(UUId)
   }
 
   async findById(id) {
-    return await this.casoUso.findById(id)
+    return await this.repository.findById(id)
   }
 
   async insert (dados) {
-    return await this.casoUso.insert(dados)
+    if (!dados.UUId) {
+      dados.UUId = UUIDGenerator.getUUIDV4()
+    }
+    return await this.repository.insert(dados)
   }
 
   async update (dados) {
-    return await this.casoUso.update(dados)
+    return await this.repository.update(dados)
   }
 
   async delete (UUId) {
-    return await this.casoUso.delete(UUId)
+    return await this.repository.delete(UUId)
   }
 }
