+++ src/app/route/produtoRoute.js
@@ -3,25 +3,25 @@ module.exports = function (app) {
   var util = new app.src.app.server.Util(app)
 
   app.get('/produto', async function (req, res) {
-    const entitys = await controller.index(req.body.attributes, req.body.filter, req.body.order)
+    const entities = await controller.index(req.body.attributes, req.body.filter, req.body.order)
     
-    if (entitys.length > 0 ) {
-      res.status(200).json(entitys)
+    if (entities.length > 0 ) {
+      res.status(200).json(entities)
     } else {
       res.status(400).json(util.montarMensagemJson(
-        'Recurso não encontrado. ' + entitys.length + ' registros encontrados.'
+        'Recurso não encontrado. ' + entities.length + ' registros encontrados.'
       ))
     }
   })
 
   app.get('/produto/paginacao', async function (req, res) {
-    const pageEntitys = await controller.findAndPaginate(req.body.attributes, req.body.filter, req.body.order, req.body.page)
+    const pageEntities = await controller.findAndPaginate(req.body.attributes, req.body.filter, req.body.order, req.body.page)
     
-    if (pageEntitys.rows.length > 0 ) {
-      res.status(200).json(pageEntitys)
+    if (pageEntities.rows.length > 0 ) {
+      res.status(200).json(pageEntities)
     } else {
       res.status(400).json(util.montarMensagemJson(
-        'Recurso não encontrado. ' + pageEntitys.rows.length + ' registros encontrados.'
+        'Recurso não encontrado. ' + pageEntities.rows.length + ' registros encontrados.'
       ))
     }
   })
